+++ src/scripts/juliaprojectcreatescripts/create_ls_project.jl
@@ -2,6 +2,7 @@ using Pkg
 
 packages_to_dev = [
     PackageSpec(path="../../../packages/CSTParser"),
+    PackageSpec(path="../../../packages/Combinatorics"),
     PackageSpec(path="../../../packages/CommonMark"),
     PackageSpec(path="../../../packages/DataStructures"),
     PackageSpec(path="../../../packages/Glob"),