+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 * Setting id for runtime inlay hints changed to `julia.inlayHints.runtime.enabled` ([#3519](https://github.com/julia-vscode/julia-vscode/pull/3519))
 
+### Fixed
+* Only update notebook metadata when it has actually changed ([#3530](https://github.com/julia-vscode/julia-vscode/pull/3530))
+
 ## [1.66.0] - 2024-01-09
 ### Changed
 * Default of `julia.persistentSession.closeStrategy` changed to overridable ([#3494](https://github.com/julia-vscode/julia-vscode/pull/3494))